Node DataToken cannot be set to a draft version
Regression after #64. When setting the token to a draft version
it is not possible to render the page (access denied).
But the token is needed because otherwise it is always the StartPage
that is set in currentContentContext.Content in ContentPropertiesStack.
This fix will make sure that the Authorize filter will not be used when
the request is in external review context.

Closes #71
CLoses #74
Closes #53