Update documentSize once the iframe is loaded/resized
Additionally we have to bind resize once the iframe is loaded,
otherwise once the context is changed and new url is loaded into the
iframe we won't be attaching to the new contentWindow object.
Now we will attach to resize once the iframe document is ready,
additionally we have to attach to beforeunload so that we can cleanup
old listener when switching context.
We also should read the offsetHeight & offsetWidth from the
actual document loaded by the iframe, not the iframe element.
That caused issues when the page content is higher than the iframe
element.