Append token at the end of a real url
In order to be able to render elements that use CurrentPage
property we should try to use a public url if possible.
If a page has been published at least once then we can just
append the token to it.
The start page will only be used when generating tokens for
pages that has never been published before in order to bypass
the authorization filter.