Add ability to add tokens to projects
When resolving the links on the page we would pass the token
to them so that a once generated link can be used to click
through the whole project easily.
We assume that if a user was given a link with access to a page
from a project then he should be able to see all other
project items.

Closes #42