Set up resize and beforeunload events on the first render
Once the component is initialized for the first time, the
iframe is already there on the page.
The extra listener to the 'load' event is only to handle
scenario when clicking on the page tree and changing
context while the ReviewWidget instance has already been
created.
If the widget is already created then we will just
load new pins from the store.
ShowReviews is no longer a toggle button, we always reset it
to false after every context change so there is no point
in having two contextchanged event handlers.