Improve ioc registration order
When intercepting default IContentLoader we need to make sure to
fall back to it in some cases but without the risk of running
into an infinite loop. The default implementation should be passed
in the constructor to the wrapper, named `defaultContentLoader` so
that it's clear it will not go through the interceptor once again.

Closes #255